refactor drawer menu
